add component for home screen
